+++ tests/on-device/config-flake-hosts-localhost.cfg.nix
@@ -1,0 +1,7 @@
+{ pkgs, ... }:
+
+{
+  system.stateVersion = "23.05";
+
+  networking.hosts."127.0.0.2" = [ "localhost" ];
+}